enhance: Semantic Scholar
