+++ camel/toolkits/__init__.py
@@ -45,7 +45,7 @@ from .human_toolkit import HumanToolkit
 from .stripe_toolkit import StripeToolkit
 from .video_toolkit import VideoDownloaderToolkit
 from .dappier_toolkit import DappierToolkit
-from .semanticscholar_toolkit import SemanticScholarToolkit
+from .semantic_scholar_toolkit import SemanticScholarToolkit
 
 __all__ = [
     'BaseToolkit',
